Move observable related classes to a package
There is a clear split between observable and scheduler related code.
