+++ src/main/kotlin/com/kimarilem/bukkit/rxbukkit/internal/observable/BukkitEventExecutor.kt
@@ -1,4 +1,4 @@
-package com.kimarilem.bukkit.rxbukkit.internal
+package com.kimarilem.bukkit.rxbukkit.internal.observable
 
 import io.reactivex.Emitter
 import org.bukkit.event.Event
